+++ Objects/Puzzle.cpp
@@ -63,7 +63,7 @@ std::ostream& operator<<(std::ostream& os, const Puzzle& obj) {
     }
     os << "\n";
     os << "User Answer: ";
-    for (int i = 0; i < obj.userAnswer.size(); i++) {
+    for (unsigned long int i = 0; i < obj.userAnswer.size(); i++) {
         os << obj.userAnswer[i] << " ";
     }
     os << "\n";
